% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{alg.A1}
\alias{alg.A1}
\title{PB test for main effect of one-way ANOVA}
\usage{
alg.A1(ns, ybars, s2, a,L)
}
\arguments{
\item{ns}{sample size for each group}

\item{ybars}{sample mean for each group}

\item{s2}{sample variance for each group}

\item{a}{level of the factor}

\item{L}{Number of simulated values for the distribution}
}
\value{
the simulated p-value
}
\description{
Using Parametric Bootstrap to simulate a distribution for the main effect of one-way ANOVA
}
\examples{

#see Q.Amc_oneway


}
