% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{alg.BC}
\alias{alg.BC}
\title{test two factor interaction}
\usage{
alg.BC(ns, ybars, s2, a, b, c, L)
}
\arguments{
\item{ns}{sample size for each group}

\item{ybars}{sample mean for each group}

\item{s2}{sample variance for each group}

\item{a}{Number of levels for factor A}

\item{b}{Number of levels for factor B}

\item{c}{Number of levels for factor C}

\item{L}{Number of simulated values for the distribution}
}
\value{
Q:	p_value for the two factor interaction test
}
\description{
Using Parametric Bootstrap to simulate a distribution and find a p-value for the test.
}
\examples{

#See Q.ABmc
# note that the ns, ybars and s2 vectors need to be in the order reflecting subscripts

# 111, 112, 113...., 121, 122, 123, ... , ... abc. The summarySE function from the package

# Rmisc is handy for doing this.  The order the user specifies the "groupvars" argument will

# put the factors in order A, B, C.  This order will matter when testing different two-way

# interaction terms and different main effects.  See comments in the potato example.

}
