% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{alg.C.AB}
\alias{alg.C.AB}
\title{tests factor C when only AB interaction is present.}
\usage{
alg.C.AB(ns, ybars, s2, a, b, c, L)
}
\arguments{
\item{ns}{sample size for each group}

\item{ybars}{sample mean for each group}

\item{s2}{sample variance for each group}

\item{a}{Number of levels for factor A}

\item{b}{Number of levels for factor B}

\item{c}{Number of levels for factor C}

\item{L}{Number of simulated values for the distribution}
}
\value{
Q:	p_value for the test for factor C main effect when only AB interaction is present.
}
\description{
Using Parametric Bootstrap to simulate a distribution and find a p-value for the test
}
\examples{

#See Q.ABmc




}
