% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{alg.C}
\alias{alg.C}
\title{for three-way ANOVA that has no significant interaction terms to test main effects}
\usage{
alg.C(ns, ybars, s2, a, b, c, L)
}
\arguments{
\item{ns}{sample size for each group}

\item{ybars}{sample mean for each group}

\item{s2}{sample variance for each group}

\item{a}{Number of levels for factor A}

\item{b}{Number of levels for factor B}

\item{c}{Number of levels for factor C}

\item{L}{Number of simulated values for the distribution}
}
\value{
a simulated p-value for testing a main effect
}
\description{
Using Parametric Bootstrap to simulate a distribution and find a p-value for the test
}
\examples{

#See Q.Amc

}
