% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barleyh20.R
\docType{data}
\name{barleyh20}
\alias{barleyh20}
\title{barleyh20 data}
\format{
This data frame contains the following columns:

\describe{
\item{genotype:}{ 1 = Troubadour, 2 = mutant from Troubadour}

\item{site:}{ 1 = Bell-lloc, Spain, 2 = Dundee, Scotland}

\item{year:}{ 91 = 1991, 92 = 1992}

\item{replicate: }{Replicate}

\item{time:}{Steeping Time (Hours)}

\item{wt}{Weight of sample (grams)}
}
}
\usage{
data(barleyh20)
}
\description{
Water uptake in barley for 2 genotypes, 2 sites, 2 years
for various periods of steeping time. 2 reps per treatment combination.
}
\references{
J.L. Molina-Cano, T. Ramo, et al. (1995). "Effect of Grain Composition
Water Uptake by Malting Barley: A Genetic and Environmental Study," Journal of
the Institute of Brewing, Vol. 101, #2, pp. 79-83.
}
\keyword{datasets}
