% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fedata.R
\docType{data}
\name{fedata}
\alias{fedata}
\title{fedata data}
\format{
This data frame contains the following columns:

\describe{
\item{depth:}{ depth of the water}

\item{Y:}{ Iron content of water }

}
}
\usage{
data(fedata)
}
\description{
Iron content of water at various water depths.
}
\references{
paper: {https://www.jstor.org/stable/1351176}
}
\keyword{datasets}
